requirements.txt
instagrapi==2.0.0
python-dotenv==1.0.0
openai==1.6.1
Pillow==10.1.0
schedule==1.2.1
python-dateutil==2.8.2 